% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame that represents the graph.}

\item{init_node}{A number that represents the starting node .}
}
\value{
A vector, the shortest path to every other node.
}
\description{
Dijkstra Algorithm
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
