% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greates common divisior of two numbers}
\usage{
euclidean(n1, n2)
}
\arguments{
\item{n1}{A number.}

\item{n2}{A number.}
}
\value{
The greatest common divisior of \code{n1} and \code{n2}.
}
\description{
Find the greates common divisior of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean algorithm.
}
