\name{Rlab3-package}
\alias{Rlab3-package}
\alias{Rlab3}
\docType{package}
\title{
\packageTitle{Rlab3}
}
\description{
\packageDescription{Rlab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rlab3}
\packageIndices{Rlab3}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{Rlab3}

Maintainer: \packageMaintainer{Rlab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
