% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean division}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number, the Greatest Common Divisor of \code{a} and \code{b}.
}
\description{
Euclidean alorithm finds the Greatest Common Divisor of two integers
}
\examples{
euclidean(123612,13892347912)
euclidean(100,1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
