% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{A data frame with three variables (v1, v2 and w).
\describe{
  \item{v1}{the first edge of the graph}
  \item{v2}{the second edge of the graph}
  \item{w}{the weight of the edge}
}}
\usage{
wiki_graph
}
\description{
This dataset contains three variables (v1, v2 and w) which are the edges 
of the graph (from v1 to v2) with the weight of the edge (w).
}
\references{
\url{https://en.wikipedia.org/wiki/Graph#Mathematics}
}
\keyword{datasets}
